Reorganizar estructura según el patrón Repository
