--- lib/core/README.md
@@ -1,3 +1,0 @@
-# Core
-
-Esta carpeta contiene todo lo que es **compartido** entre las diferentes features de la aplicación.
