--- lib/core/database/README.md
@@ -1,3 +1,0 @@
-# database
-
-Aquí irán los servicios relacionados con la configuración y acceso a la base de datos (por ejemplo, SQLite o Hive).
