--- lib/core/error/README.md
@@ -1,3 +1,0 @@
-# error
-
-ontendrá la lógica para manejar errores y excepciones globales.
