--- lib/core/network/README.md
@@ -1,3 +1,0 @@
-# network
-
-Aquí estarán los clientes y configuraciones para conectarse a APIs externas.
