--- lib/core/theme/README.md
@@ -1,3 +1,0 @@
-# theme
-
-Aquí configuramos los estilos globales de la aplicación (como colores y fuentes).
