--- lib/core/utils/README.md
@@ -1,3 +1,0 @@
-# utils
-
-Aquí colocamos funciones, constantes, y helpers reutilizables.
