--- lib/core/widgets/README.md
@@ -1,3 +1,0 @@
-# widgets
-
-Aquí colocamos widgets reutilizables entre varias features.
