--- lib/features/README.md
@@ -1,3 +1,0 @@
-# Features
-
-Esta carpeta agrupa las diferentes funcionalidades o módulos de la aplicación. Cada funcionalidad tiene su propia estructura interna basada en el patrón Clean Architecture.
