+++ lib/features/dashboard/domain/entities/client_model.dart
