+++ lib/features/dashboard/presentation/pages/detalle_tarima_screen.dart
