+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'features/dashboard/widgets/dashboard_widget.dart'; // Importamos el archivo del Dashboard desde su nueva ubicación
+import 'features/dashboard/presentation/widgets/dashboard_widget.dart'; // Importamos el archivo del Dashboard desde su nueva ubicación
 
 void main() {
   runApp(const MainApp());
